refactor: remove retry utilities from `mpc-utils`
