+++ chain-signatures/utils/src/lib.rs
@@ -1,6 +1,5 @@
-//! Generic shared utilities for the MPC workspace (time, task, retry).
+//! Generic shared utilities for the MPC workspace (time, task).
 //! Domain-specific helpers live in `mpc-chain-integration-core`.
 
-pub mod retry;
 pub mod task;
 pub mod time;
